Extract non-workPackage specific logic to a separate service to be able to use it other calendars as well (e.g spent time graph)
