+++ frontend/src/app/features/calendar/wp-calendar/wp-calendar.component.ts
@@ -33,22 +33,24 @@ import { debounceTime } from 'rxjs/operators';
 import { ConfigurationService } from 'core-app/core/config/configuration.service';
 import { UntilDestroyedMixin } from 'core-app/shared/helpers/angular/until-destroyed.mixin';
 import { SchemaCacheService } from 'core-app/core/schemas/schema-cache.service';
-import {
-  CalendarViewEvent,
-  OpCalendarService,
-} from 'core-app/features/calendar/op-calendar.service';
 import { Subject } from 'rxjs';
 import { CurrentProjectService } from 'core-app/core/current-project/current-project.service';
 import interactionPlugin, { EventResizeDoneArg } from '@fullcalendar/interaction';
 import { HalResourceEditingService } from 'core-app/shared/components/fields/edit/services/hal-resource-editing.service';
 import { HalResourceNotificationService } from 'core-app/features/hal/services/hal-resource-notification.service';
 import { splitViewRoute } from 'core-app/features/work-packages/routing/split-view-routes.helper';
+import {
+  CalendarViewEvent,
+  OpWorkPackagesCalendarService,
+} from 'core-app/features/calendar/op-work-packages-calendar.service';
+import { OpCalendarService } from 'core-app/features/calendar/op-calendar.service';
 
 @Component({
   templateUrl: './wp-calendar.template.html',
   styleUrls: ['./wp-calendar.sass'],
   selector: 'op-wp-calendar',
   providers: [
+    OpWorkPackagesCalendarService,
     OpCalendarService,
   ],
 })
@@ -80,6 +82,7 @@ export class WorkPackagesCalendarComponent extends UntilDestroyedMixin implement
     private sanitizer:DomSanitizer,
     private configuration:ConfigurationService,
     readonly calendar:OpCalendarService,
+    readonly workPackagesCalendar:OpWorkPackagesCalendarService,
     readonly currentProject:CurrentProjectService,
     readonly halEditing:HalResourceEditingService,
     readonly halNotification:HalResourceNotificationService,
@@ -115,7 +118,7 @@ export class WorkPackagesCalendarComponent extends UntilDestroyedMixin implement
       successCallback(events);
     } else {
       this
-        .calendar
+        .workPackagesCalendar
         .currentWorkPackages$
         .subscribe((collection:WorkPackageCollectionResource) => {
           const events = this.updateResults((collection));
@@ -123,7 +126,7 @@ export class WorkPackagesCalendarComponent extends UntilDestroyedMixin implement
         });
     }
 
-    void this.calendar.updateTimeframe(fetchInfo, this.currentProject.identifier || undefined);
+    void this.workPackagesCalendar.updateTimeframe(fetchInfo, this.currentProject.identifier || undefined);
   }
 
   // eslint-disable-next-line @angular-eslint/use-lifecycle-interface
@@ -159,7 +162,7 @@ export class WorkPackagesCalendarComponent extends UntilDestroyedMixin implement
         // Currently the calendar widget is shown on multiple pages,
         // but only the calendar module itself is a partitioned query space which can deal with a split screen request
         if (this.$state.includes('calendar')) {
-          this.calendar.openSplitView(workPackageId);
+          this.workPackagesCalendar.openSplitView(workPackageId);
         } else {
           void this.$state.go(
             'work-packages.show',
@@ -176,7 +179,7 @@ export class WorkPackagesCalendarComponent extends UntilDestroyedMixin implement
     void this.configuration.initialized
       .then(() => {
         this.calendarOptions$.next(
-          this.calendar.calendarOptions(additionalOptions),
+          this.workPackagesCalendar.calendarOptions(additionalOptions),
         );
       });
   }
@@ -199,32 +202,32 @@ export class WorkPackagesCalendarComponent extends UntilDestroyedMixin implement
     }
 
     const workPackageId = eventContainer.dataset.workPackageId as string;
-    this.calendar.showEventContextMenu({ workPackageId, event });
+    this.workPackagesCalendar.showEventContextMenu({ workPackageId, event });
   }
 
   private setupWorkPackagesListener():void {
-    this.calendar.workPackagesListener$(() => {
+    this.workPackagesCalendar.workPackagesListener$(() => {
       this.alreadyLoaded = true;
       this.ucCalendar.getApi().refetchEvents();
     });
   }
 
   private updateResults(collection:WorkPackageCollectionResource) {
-    this.calendar.warnOnTooManyResults(collection, this.static);
+    this.workPackagesCalendar.warnOnTooManyResults(collection, this.static);
     return this.mapToCalendarEvents(collection.elements);
   }
 
   private mapToCalendarEvents(workPackages:WorkPackageResource[]) {
     const events = workPackages.map((workPackage:WorkPackageResource) => {
-      const startDate = this.calendar.eventDate(workPackage, 'start');
-      const endDate = this.calendar.eventDate(workPackage, 'due');
+      const startDate = this.workPackagesCalendar.eventDate(workPackage, 'start');
+      const endDate = this.workPackagesCalendar.eventDate(workPackage, 'due');
 
       const exclusiveEnd = moment(endDate).add(1, 'days').format('YYYY-MM-DD');
 
       return {
         title: workPackage.subject,
         start: startDate,
-        editable: this.calendar.eventDurationEditable(workPackage),
+        editable: this.workPackagesCalendar.eventDurationEditable(workPackage),
         end: exclusiveEnd,
         allDay: true,
         className: `__hl_background_type_${workPackage.type.id}`,
@@ -240,7 +243,7 @@ export class WorkPackagesCalendarComponent extends UntilDestroyedMixin implement
   }
 
   private async updateEvent(info:EventResizeDoneArg|EventDropArg):Promise<void> {
-    const changeset = this.calendar.updateDates(info);
+    const changeset = this.workPackagesCalendar.updateDates(info);
 
     try {
       const result = await this.halEditing.save(changeset);
@@ -254,7 +257,7 @@ export class WorkPackagesCalendarComponent extends UntilDestroyedMixin implement
   private handleDateClicked(info:DateSelectArg) {
     const defaults = {
       startDate: info.startStr,
-      dueDate: this.calendar.getEndDateFromTimestamp(info.end),
+      dueDate: this.workPackagesCalendar.getEndDateFromTimestamp(info.end),
     };
 
     void this.$state.go(
