+++ frontend/src/app/features/calendar/wp-calendar/wp-calendar.template.html
@@ -13,6 +13,6 @@
     </ng-container>
     <div
         *ngIf="static"
-        [textContent]="calendar.tooManyResultsText"
+        [textContent]="workPackagesCalendar.tooManyResultsText"
         class="op-wp-calendar--notification"></div>
 </div>
