+++ frontend/src/app/features/team-planner/team-planner/add-work-packages/add-existing-pane.component.html
@@ -30,7 +30,7 @@
           class="op-add-existing-pane--wp"
           [attr.data-qa-selector]="'op-add-existing-pane--wp-' + wp.id"
           (stateLinkClicked)="openStateLink($event)"
-          (cardClicked)="calendar.onCardClicked($event)"
+          (cardClicked)="workPackagesCalendar.onCardClicked($event)"
         ></wp-single-card>
       </ng-container>
     </div>
