+++ frontend/src/app/features/team-planner/team-planner/calendar-drag-drop.service.ts
@@ -11,6 +11,7 @@ import { SchemaCacheService } from 'core-app/core/schemas/schema-cache.service';
 import { AuthorisationService } from 'core-app/core/model-auth/model-auth.service';
 import { I18nService } from 'core-app/core/i18n/i18n.service';
 import { OpCalendarService } from 'core-app/features/calendar/op-calendar.service';
+import { OpWorkPackagesCalendarService } from 'core-app/features/calendar/op-work-packages-calendar.service';
 
 @Injectable()
 export class CalendarDragDropService {
@@ -30,7 +31,7 @@ export class CalendarDragDropService {
   constructor(
     readonly authorisation:AuthorisationService,
     readonly schemaCache:SchemaCacheService,
-    readonly calendarService:OpCalendarService,
+    readonly workPackagesCalendarService:OpWorkPackagesCalendarService,
     readonly I18n:I18nService,
   ) {
   }
@@ -98,7 +99,7 @@ export class CalendarDragDropService {
       return { disabled: true, reason: this.text.draggingDisabled.permissionDenied };
     }
 
-    if (!this.calendarService.dateEditable(workPackage)) {
+    if (!this.workPackagesCalendarService.dateEditable(workPackage)) {
       return { disabled: true, reason: this.text.draggingDisabled.fallback };
     }
