+++ frontend/src/app/features/team-planner/team-planner/planner/team-planner.component.ts
@@ -85,6 +85,7 @@ import { CapabilitiesResourceService } from 'core-app/core/state/capabilities/ca
 import { ICapability } from 'core-app/core/state/capabilities/capability.model';
 import { ToastService } from 'core-app/shared/components/toaster/toast.service';
 import { LoadingIndicatorService } from 'core-app/core/loading-indicator/loading-indicator.service';
+import { OpWorkPackagesCalendarService } from 'core-app/features/calendar/op-work-packages-calendar.service';
 
 @Component({
   selector: 'op-team-planner',
@@ -142,7 +143,7 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
       filter((dragging) => !!dragging),
       map((dragging) => {
         const workPackage = (dragging as EventDragStartArg).event.extendedProps.workPackage as WorkPackageResource;
-        const dateEditable = this.calendar.dateEditable(workPackage);
+        const dateEditable = this.workPackagesCalendar.dateEditable(workPackage);
         const resourceEditable = this.eventResourceEditable(workPackage);
         return dateEditable && resourceEditable;
       }),
@@ -284,6 +285,7 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
     private I18n:I18nService,
     readonly injector:Injector,
     readonly calendar:OpCalendarService,
+    readonly workPackagesCalendar:OpWorkPackagesCalendarService,
     readonly halEditing:HalResourceEditingService,
     readonly halNotification:HalResourceNotificationService,
     readonly schemaCache:SchemaCacheService,
@@ -381,13 +383,13 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
     void this.configuration.initialized
       .then(() => {
         this.calendarOptions$.next(
-          this.calendar.calendarOptions({
+          this.workPackagesCalendar.calendarOptions({
             schedulerLicenseKey: 'GPL-My-Project-Is-Open-Source',
             selectable: true,
             plugins: [resourceTimelinePlugin, interactionPlugin],
             titleFormat: { year: 'numeric', month: 'long', day: 'numeric' },
             buttonText: { today: this.text.today },
-            initialView: this.calendar.initialView || 'resourceTimelineWeek',
+            initialView: this.workPackagesCalendar.initialView || 'resourceTimelineWeek',
             headerToolbar: {
               left: '',
               center: 'title',
@@ -511,7 +513,7 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
     failureCallback:(error:unknown) => void,
   ):void|PromiseLike<EventInput[]> {
     this
-      .calendar
+      .workPackagesCalendar
       .currentWorkPackages$
       .pipe(
         withLatestFrom(this.assigneeCaps$),
@@ -531,7 +533,7 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
         failureCallback,
       );
 
-    void this.calendar.updateTimeframe(fetchInfo, this.projectIdentifier);
+    void this.workPackagesCalendar.updateTimeframe(fetchInfo, this.projectIdentifier);
   }
 
   /**
@@ -685,7 +687,7 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
         }
 
         const assignee = this.wpAssignee(workPackage);
-        const durationEditable = this.calendar.eventDurationEditable(workPackage);
+        const durationEditable = this.workPackagesCalendar.eventDurationEditable(workPackage);
         const resourceEditable = this.eventResourceEditable(workPackage);
 
         return {
@@ -711,7 +713,7 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
     this.openNewSplitCreate(
       info.startStr,
       // end date is exclusive
-      this.calendar.getEndDateFromTimestamp(info.end),
+      this.workPackagesCalendar.getEndDateFromTimestamp(info.end),
       info.resource?.id || '',
     );
   }
@@ -762,7 +764,7 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
   }
 
   private async updateEvent(info:EventResizeDoneArg|EventDropArg|EventReceiveArg):Promise<void> {
-    const changeset = this.calendar.updateDates(info);
+    const changeset = this.workPackagesCalendar.updateDates(info);
 
     const resource = info.event.getResources()[0];
     if (resource) {
@@ -803,7 +805,7 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
   ):{ [key:string]:string|string[] } {
     const constraints:{ [key:string]:string|string[] } = {};
 
-    if (!this.calendar.eventDurationEditable(wp) && !wp.date) {
+    if (!this.workPackagesCalendar.eventDurationEditable(wp) && !wp.date) {
       constraints.start = this.wpStartDate(wp);
       constraints.end = this.wpEndDate(wp);
     }
@@ -822,11 +824,11 @@ export class TeamPlannerComponent extends UntilDestroyedMixin implements OnInit,
   }
 
   private wpStartDate(wp:WorkPackageResource):string {
-    return this.calendar.eventDate(wp, 'start');
+    return this.workPackagesCalendar.eventDate(wp, 'start');
   }
 
   private wpEndDate(wp:WorkPackageResource):string {
-    const endDate = this.calendar.eventDate(wp, 'due');
+    const endDate = this.workPackagesCalendar.eventDate(wp, 'due');
     return moment(endDate).add(1, 'days').format('YYYY-MM-DD');
   }
